update time of data calculation

+++ app/src/main/java/com/mmd/cityweather/common/data/api/model/CurrentWeatherApi.kt
@@ -15,7 +15,7 @@ data class CurrentWeatherApi(
     @Json(name = "coord")
     val coord: Coord?,
     @Json(name = "dt")
-    val dt: Int?,
+    val dt: Long?,
     @Json(name = "id")
     val id: Long?,
     @Json(name = "main")
